+++ src/pages/index.js
@@ -1,65 +1,9 @@
 import { graphql, Link, navigate } from "gatsby"
 import React, { useState } from "react"
 import { Router } from "@reach/router"
-import SheetPage from "../util/sheet-details"
-import { match } from "../util/util"
-import * as styles from '../styles/home.module.css'
-import Album from '../../assets/album.svg'
-import Wohldenberg from '../../assets/wohldenberg.svg'
-import Ukulele from '../../assets/ukulele.svg'
-import Magnifier from '../../assets/magnifier.svg'
+import SheetPage from "../components/sheet-details"
+import Home from "../components/home"
 
-const Tags = ({tags}) => {
-  const allTags = tags.replace(", ", ",").split(",")
-  return (<span className={styles.tags}>
-  {allTags.map(tag => {
-    if (tag.startsWith("album:")) {
-      return (<><Album className={styles.icon} />{tag.replace("album: ", "").replace("album:", "")}</>)
-    }
-    if (tag === "wohldenberg") {
-      return (<Wohldenberg className={styles.icon} />)
-    }
-    if (tag === "ukulele") {
-      return (<Ukulele className={styles.icon} />)
-    }
-    return (<>{tag}</>)
-  })}
-  </span>)
-}
-
-const Home = ({sheetlist, searchString, setSearchString}) => {
-  if (searchString !== "") {
-    sheetlist.sort((a,b) => {return match(b.frontmatter, searchString) - match(a.frontmatter, searchString)})
-  }
-  return (
-    <>
-      <div className={styles.searchbar}>
-        <Magnifier className={styles.icon} />
-        <input onChange={(event) => {setSearchString(event.target.value)}} />
-      </div>
-      <h1 style={{textAlign: "center", margin: "0px"}}>Delyrium</h1>
-      <div className={styles.cardbox}>
-        {sheetlist.map(sheet => {
-          if (searchString === "" || match(sheet.frontmatter, searchString) >= 0.5) {
-            return (
-              <Link to={`/sheets/${sheet.frontmatter.slug}`} key={sheet.id} className={styles.card}>
-                <div>
-                  <h3 className={styles.title}>{sheet.frontmatter.title}</h3>
-                  <div className={styles.band}>
-                    {sheet.frontmatter.band}
-                    <Tags tags={sheet.frontmatter.tags ?? ""} />
-                  </div>
-                </div>
-              </Link>
-            )
-          } else {
-            return (<></>)
-          }
-        })}
-      </div>
-    </>
-  )
-}
 
 // TODO: Loading all pages at once can perform badly for huge libraries, fix with LazyLoad vs offline mode?
 export default function Sheetlist({ data }) {
